requirements.txt
certifi==2019.11.28
chardet==3.0.4
DateTime==4.3
httplib2==0.14.0
icalendar==4.0.4
icalevents==0.1.23
idna==2.8
pbr==5.4.4
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
stevedore==1.31.0
urllib3==1.25.7
virtualenv==16.7.8
virtualenv-clone==0.5.3
zope.interface==4.7.1
